working at school

+++ src/algoritmoGenetico/Operaciones.java
@@ -5,7 +5,7 @@ public class Operaciones {
 		return a + (int)(Math.random() * ((b - a) + 1));		
 	}
 	
-	public static int aleatorioEntreExcepto(int a, int b, byte[] fijos){
+	public static int aleatorioEntreExcepto(int a, int b, int[] fijos){
 		boolean[] mascara = new boolean[9];
 		for (int i=0; i<9; i++){
 			if (fijos[i]!=0){
@@ -15,7 +15,7 @@ public class Operaciones {
 		int random=0;
 		boolean randomValido = false;
 		while(!randomValido) {
-			random = (byte)Operaciones.aleatorioEntre(a, b);
+			random = (int)Operaciones.aleatorioEntre(a, b);
 			if (!mascara[random - 1]){
 				randomValido=true;
 				mascara[random - 1]=true;
